refac: slack config refac
