--- src/main/kotlin/com/oksusu/susu/slack/config/SlackConfig.kt
@@ -1,10 +1,0 @@
-package com.oksusu.susu.slack.config
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@EnableConfigurationProperties(
-    SlackAlarmConfig::class
-)
-class SlackConfig
